added experimental literal resolver
Change AST literal integer types to i128/u128 and update parser casts to match. Add a new hir_literal_interpreter crate (Cargo.toml, .gitignore, lock, and src/) implementing literal resolution and interpretation (unary, binary, and utils) to enable constant literal evaluation. Refactor Hir ExpressionKind to use dedicated Unary and Binary structs and propagate those changes across the hir parser, typed context, mir parser, tester, and displayer. Rename infer_types to infer_hir_types and update call sites. Introduce MIR conditional lowering module and reorganize expression parsing into expressions/mod.rs. Update soul_tester to integrate the interpreter, adjust compilation flow, token handling, and error reporting. Add SementicFault::debug helper. Various minor fixes to keep types and APIs consistent.

+++ soul_ast/ast_model/src/ast/literal.rs
@@ -5,8 +5,8 @@ use soul_utils::soul_names::{PrimitiveTypes, TypeModifier};
 /// A literal value in the Soul language.
 #[derive(Clone, PartialEq, serde::Serialize, serde::Deserialize)]
 pub enum Literal {
-    Int(i64),
-    Uint(u64),
+    Int(i128),
+    Uint(u128),
     Float(f64),
     Bool(bool),
     Char(char),
